// src/server/src/common/general.enum.ts
export enum formErrorValues {
  createPasswordError = 'Password should be from 8 to 30 characters',
  enterPasswordError = 'Your password might be missing some characters.',
  firstNameError = 'First name should be 3 characters at least',
  lastNameError = 'Last name should be 3 characters at least',
  topicError = 'Topic name should be from 2 to 20 characters',
}

export enum formRequireValues {
  nameMin = 3,
  nameMax = 30,
  passwordMin = 8,
  passwordax = 30,
  topicMin = 2,
  topicMax = 20,
}